[feature] : Window Event on Windows

+++ core/include/core/hash/Fnv1a.h
@@ -19,7 +19,7 @@ namespace sky {
     inline constexpr uint32_t Fnv1a32(std::string_view str)
     {
         uint32_t res = FNVOffsetBias32;
-        uint32_t size = str.size();
+        uint32_t size = static_cast<uint32_t>(str.size());
         for (uint32_t i = 0; i < size; ++i) {
             res ^= str[i];
             res *= FNVPrime32;
